\name{SimpleEx}
 \docType{data}
 \alias{SimpleEx}
 \title{Simple Data Example With Four Variables}
 \description{
   This data set gives a simple example consisting of observed data, ratio edits and a (simple) balance edit.
 }
 \usage{data(SimpleEx)}
 \format{A list containing the followings:
 \itemize{
 \item \bold{D.obs}: \code{data.frame} of observed data with \eqn{n=1000} records and \eqn{p=4} variables.
 \item \bold{Ratio.edit}: \code{data.frame} of two ratio edits.
 \item \bold{Balance.edit}: \code{data.frame} of a balance edit.
 }
 }
\seealso{\code{\link{readData}}} 
 \keyword{datasets}
